# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: 车辆信息展示系统 HTTP 契约
  version: 1.2.0
  summary: F-001 / F-003 / F-004 首次落约
  description: |
    **first release，not breaking。** `contracts/` 此前为空，本次全部为新增。

    覆盖已裁决 feature：F-001 车辆信息管理、F-003 图片与评估报告、F-004 价格记录。
    不覆盖仍为 draft 的 F-002（前台页面结构与展示文案）、F-005（回收站）、F-006（登录与账号）。

    ## 鉴权
    `/admin/*` 均声明 `security: [adminAuth: []]`。鉴权由 F-006 定稿，本期假设已登录管理员。
    本文件不规定登录方式、token 形态与刷新。未携带有效管理员凭据时返回 401 UNAUTHORIZED。
    `/public/*` 无登录。前台无用户体系。

    ## 错误
    所有错误响应体：
    `{ "code": "VEHICLE_NOT_FOUND", "message": "...", "details": {} }`
    码表与 HTTP 映射见 `errors.yaml`。公开端对不存在或未上架车辆一律 404 VEHICLE_NOT_FOUND，禁止 403。

    ## 内部方法（不是 HTTP）
    F-001 发布反查不得走本文件 paths：
    - `countImages(vehicleId) -> int`（F-003）
    - `isPriceFilled(vehicleId) -> boolean`（F-004）
    详见 `internal.yaml`。

    ## 字段与范围
    JSON 字段 camelCase。车辆资源不含价格。公开响应不得出现 vin / vinMasked。
    不做：导出、批量改价、回收站、登录注册、软删除字段、前台文案「面议」。

    ## 图片确认策略
    预签名后，确认落库为一次一张（`adminConfirmVehicleImage` / `adminConfirmVehicleReport`），
    由客户端循环。不使用 HTTP 207。部分失败由客户端按单张结果处理。
    同一 objectKey 重复确认返回已落库资源（200），不创建第二条。
servers:
  - url: http://localhost:8080
    description: 本地开发。路径无额外前缀，`/admin` 与 `/public` 位于服务根。
  - url: https://{host}
    description: 部署环境占位。前台域名备案是部署前置条件（ADR-011），host 本期不写死。
    variables:
      host:
        default: api.example.invalid
        description: 服务端主机名，上线时替换
...
security: []
paths:
...
  /admin/accounts:
    post:
      operationId: adminCreateAccount
      summary: 新建管理员账号（仅超管）
      description: |
        仅 super_admin 可调用。填 loginName + role，系统生成随机初始口令，仅在本次响应体里
        明文返回一次（不落库明文、不可再查看）。新账号 mustChangePassword=true。
        loginName 全局唯一，创建后不可修改；已停用账号的 loginName 永久占用不可复用。

        **不提供 Idempotency-Key**：与 F-001 建车辆不同，若网络超时导致客户端未收到这次响应，
        账号在服务端已创建成功但客户端拿不到那个一次性口令——这种情况**接受**（ADR-026 已裁决
        流程），恢复路径是超管再调用一次 `POST /admin/accounts/{accountId}/password-reset`
        重新生成一个新口令，不做"重放返回同一口令"的幂等设计（会破坏"口令不落库明文、不可再查"
        的安全承诺）。
        **网关/日志系统要求**：本接口响应体含明文口令，网关、APM、错误监控、反向代理日志
        **禁止记录本接口与 password-reset 接口的响应体**，仅记录状态码与耗时。
      security: [{ adminAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [loginName, role]
              properties:
                loginName: { type: string, minLength: 1, maxLength: 50 }
                role: { type: string, enum: [admin, super_admin] }
      responses:
        "201":
          description: 创建成功，含一次性明文初始口令
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminAccountCreatedResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { description: 非超管调用（FORBIDDEN）, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "409": { description: loginName 已被占用（含已停用账号）, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "500": { $ref: '#/components/responses/InternalError' }

    get:
      operationId: adminListAccounts
      summary: 账号列表（仅超管）
      description: |
        默认展示全部账号（含已停用），带角色与启用状态标签。仅 super_admin 可调用。
        默认排序：id 升序（创建顺序）。
      security: [{ adminAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        "200": { description: 账号列表, content: { application/json: { schema: { $ref: '#/components/schemas/AdminAccountList' } } } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { description: 非超管调用（FORBIDDEN）, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "500": { $ref: '#/components/responses/InternalError' }
  /admin/accounts/{accountId}/status:
    post:
      operationId: adminSetAccountStatus
      summary: 停用 / 启用账号（仅超管）
      description: |
        body.enabled=false 为停用，true 为启用。仅 super_admin 可调用。
        用单一端点 + 布尔值而非 F-001 式的独立动词端点（publish/unpublish），是因为启用/停用只有
        两态、无中间态、无前置校验，用布尔值更省；F-001 的多态转换更适合动词端点，两种风格在本
        仓库内并存是刻意选择，不是遗漏。

        检查顺序（先后如下，前面命中即短路）：
        1. 非 super_admin 调用 → 403 FORBIDDEN
        2. 目标账号不存在 → 404
        3. 目标账号是调用者自己 → 403 CANNOT_DISABLE_SELF（无论 enabled 传 true 还是 false，
           只要目标是自己就拒绝，简化规则，不必区分"停用自己"与"启用自己"这种理论上无意义的组合）
        4. 已是目标状态（enabled 与当前一致）→ 200 幂等成功，不再往下检查
        5. enabled=false 且目标是系统里最后一个启用状态的 super_admin → 409 LAST_SUPER_ADMIN
        6. 以上均不命中 → 执行变更，停用时作废该账号全部现有会话（同步一致失效，见 securityScheme）
      security: [{ adminAuth: [] }]
      parameters:
        - name: accountId
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        "200": { description: 已生效（含幂等重复调用）, content: { application/json: { schema: { $ref: '#/components/schemas/AdminAccount' } } } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { description: 非超管调用（FORBIDDEN），或对自己调用（CANNOT_DISABLE_SELF）, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "404": { description: 账号不存在, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "409": { description: 尝试停用最后一个启用超管, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "500": { $ref: '#/components/responses/InternalError' }

    delete:
      operationId: adminDeleteAccount
      summary: 删除账号（软删除，仅超管，ADR-119）
      description: |
        仅 super_admin 可调用。软删除：账号打 deleted_at 标记，不物理清除行，不可恢复。

        前置条件：目标账号必须是"已停用"（disabled）状态，否则 409。
        "不能删除自己"不作为独立校验规则：ADR-026 已禁止"停用自己"，而删除的前置条件是
        "已停用"，两条规则组合后超管天然无法把自己降到可删状态，属于隐含防御。

        删除后该账号的 loginName 永久占用不释放，不可被新账号复用，与账号创建/停用接口
        既有的 loginName 永久占用规则一致。

        重复调用同一个已删除账号返回 404（不是幂等成功）——删除是不可逆操作，需要明确反馈。

        **实现提示**：服务端判断前置条件时应使用数据库层条件更新（如
        `UPDATE ... WHERE id=? AND status='disabled' AND deleted_at IS NULL` 判断受影响行数），
        而非先查询状态再更新的两步式判断，避免并发场景下的竞态。
      security: [{ adminAuth: [] }]
      parameters:
        - name: accountId
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        "204": { description: 删除成功，无响应体 }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { description: 非超管调用（FORBIDDEN）, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "404": { description: 账号不存在，或已被软删除, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "409": { description: 账号当前不是已停用状态, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "500": { $ref: '#/components/responses/InternalError' }
  /admin/accounts/{accountId}/role:
    post:
      operationId: adminSetAccountRole
      summary: 修改账号角色（仅超管，ADR-026 允许升降级）
      description: |
        仅 super_admin 可调用。

        检查顺序（先后如下，前面命中即短路）：
        1. 非 super_admin 调用 → 403 FORBIDDEN
        2. 目标账号不存在 → 404
        3. 已是目标角色 → 200 幂等成功，不再往下检查
        4. role=admin（降级）且目标是系统里最后一个启用状态的 super_admin → 409 LAST_SUPER_ADMIN
        5. 以上均不命中 → 执行变更（不影响会话，角色变更不强制下线，下一次鉴权即按新角色生效）
      security: [{ adminAuth: [] }]
      parameters:
        - name: accountId
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role: { type: string, enum: [admin, super_admin] }
      responses:
        "200": { description: 已生效（含幂等重复调用）, content: { application/json: { schema: { $ref: '#/components/schemas/AdminAccount' } } } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { description: 非超管调用（FORBIDDEN）, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "404": { description: 账号不存在, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "409": { description: 尝试降级最后一个启用超管, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "500": { $ref: '#/components/responses/InternalError' }
  /admin/accounts/{accountId}/password-reset:
    post:
      operationId: adminResetAccountPassword
      summary: 重置他人密码（仅超管）
      description: |
        仅 super_admin 可调用。生成新随机口令，仅在本次响应体里明文返回一次。
        重置后该账号 mustChangePassword 置为 true，且全部现有会话作废。
        同样不提供 Idempotency-Key，网络超时未收到口令时可重复调用本接口生成新口令
        （本接口天然可重复调用，语义上比建号更安全）。
        网关/日志系统同样禁止记录本接口响应体（见 adminCreateAccount 的日志脱敏要求）。
      security: [{ adminAuth: [] }]
      parameters:
        - name: accountId
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        "200": { description: 重置成功，含一次性明文新口令, content: { application/json: { schema: { $ref: '#/components/schemas/AdminAccountCreatedResponse' } } } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { description: 非超管调用（FORBIDDEN）, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "404": { description: 账号不存在, content: { application/json: { schema: { $ref: './errors.yaml#/components/schemas/ErrorResponse' } } } }
        "500": { $ref: '#/components/responses/InternalError' }
components:
  securitySchemes:
    adminAuth:
      type: apiKey
      in: cookie
      name: admin_session
      description: |
        服务端会话（ADR-027），不透明会话标识，客户端不解析。8 小时滑动过期，同账号允许多端登录。
        账号停用或改密后，该账号全部现有会话作废——吊销判定必须**同步一致读**（每次请求查会话
        存储判定是否已被吊销），不允许"内存缓存+定期刷新"引入延迟窗口，否则违反 ADR-027 的
        「停用立即生效」零容忍要求。

        Cookie 属性（安全定稿，reviewer-security 阻塞项）：`HttpOnly; Secure; SameSite=Strict;
        Path=/admin`。管理后台前端与本 API 同源部署（同域名，不同于 ADR-028 里"前台与后台
        不同域"——那是前台公开站点与后台的隔离，后台前端与后台 API 视为同一部署单元），
        因此不需要跨域 CORS credentials 配置，`SameSite=Strict` 天然覆盖大部分 CSRF 场景。

        CSRF 防护（安全定稿）：除 `SameSite=Strict` 外，全部状态变更接口（POST/PUT/DELETE）
        额外校验请求的 `Origin` header 必须等于部署域名，不一致则 403 FORBIDDEN，双重防护。

        密码存储：服务端用 **argon2id**（或 bcrypt，cost factor ≥ 12）存储密码哈希，
        **禁止**明文或 MD5/SHA256 直接摘要落库。全部 `/admin/*` 接口强制 HTTPS 传输，
        不接受明文 HTTP。密码比对使用哈希库自带的恒定时间比较，避免时序侧信道。

        未携带有效会话 → 401 UNAUTHORIZED。
        已登录但角色不足（如管理员访问超管专属接口）→ 403 FORBIDDEN。
        已登录但账号已被停用（会话吊销前的极短竞态窗口内）→ 401 UNAUTHORIZED（视为会话已失效，
        不使用 403，避免与"角色不足"混淆）。
        已登录但 mustChangePassword=true 且调用的不是改密接口本身 → 403 MUST_CHANGE_PASSWORD。
...
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      description: 页码，从 1 开始。超出实际页数时返回空 items，total 仍为匹配总数。
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: 每页条数。工程上限 100，不是车辆规模上限（ADR-007）。
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: 请求不合法（VALIDATION_ERROR / MISSING_VERSION / IDEMPOTENCY_KEY_REQUIRED / INVALID_OBJECT_KEY / INVALID_CONTENT_TYPE / FILE_TOO_LARGE / IMAGE_ORDER_MISMATCH / OBJECT_NOT_UPLOADED）
      content:
        application/json:
          schema:
            $ref: "./errors.yaml#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未登录或登录已失效
      content:
        application/json:
          schema:
            $ref: "./errors.yaml#/components/schemas/ErrorResponse"
          example:
            code: UNAUTHORIZED
            message: 未登录或登录已失效
            details: {}
...
    InternalError:
      description: 服务内部错误。发布反查失败不得降级为前置条件失败。
      content:
        application/json:
          schema:
            $ref: "./errors.yaml#/components/schemas/ErrorResponse"
          example:
            code: INTERNAL_ERROR
            message: 服务内部错误
            details: {}
...
  schemas:
...
    AdminAccount:
      type: object
      additionalProperties: false
      description: 账号资源。不含密码/密码哈希。
      required: [id, loginName, role, enabled, mustChangePassword, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64, minimum: 1 }
        loginName: { type: string }
        role: { type: string, enum: [admin, super_admin] }
        enabled: { type: boolean }
        mustChangePassword: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AdminAccountList:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AdminAccount' } }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
    AdminAccountCreatedResponse:
      type: object
      additionalProperties: false
      description: 新建账号或重置密码后的一次性响应，initialPassword 只在本次返回，不再可查。
      required: [account, initialPassword]
      properties:
        account: { $ref: '#/components/schemas/AdminAccount' }
        initialPassword:
          type: string
          description: 一次性明文初始口令，仅本次响应体携带，服务端不落库明文